use tables rather than tabsets for platform-specific python shell commands
With only a limited set of users to date, I've already seen multiple times where users "missed" that there was a tab for their platform and entered the wrong command.

While this presentation is less compact, I think it will lead to less errors. Note that windows commands (using `py`) are listed first, and those commands will actually yield an error on other platforms (causing the user to re-consult the docs). Previously we had Mac/Linux on the first tab -- those commands would actually succeed on Windows but potentially yield the wrong result (via using the wrong version of Python).