Additional option
The URLs of my blog posts slightly changed when porting my blog to Quarto, so I needed some redirects. As I use [GitHub Actions](https://quarto.org/docs/publishing/netlify.html#github-action) to deploy the rendered content to Netlify, I added a netlify.toml file with the redirects. However, this file is not included in the rendered content by default and it kept me busy a while figuring out how to achieve this. This contribution may save others some time. In general, this option is interesting for any file that is not included in the rendered content by default.