reference template partials
When searching for example for "quarto custom title page" I get the title blocks page, which explains on how to customize the blocks. I think a guide or reference on how to change specific parts of a document, one of it being the title page/block could be useful for others.

Any thoughts on it?