Added guide on how to enable snippets suggestion
I was looking for a way to enable snippet suggestions in `qmd` files and found a stackoverflow post that explained it. I can't find it anymore though, but I'd like to add a quick guide on how to enable it.
Also it seems that the Quarto extension overrides the default setting.