Specify Cross Ref identifier case constraint
i was going a little crazy as a beginner as to why my reference wasn't working. From guessing and testing it was because i had a couple of uppercase letters sprinkled in. Thought it should be added in the docs. Not sure if there are other simple rules to know but from now on ill be sticking to only hyphens, numbers, and lowercase. Cheers.