Suggest linux users put execs in ~/.local/bin
On many Linux distributions, both `~/bin` and `~/.local/bin` will
already be on the PATH, thanks to this common snippet in the
/etc/profile file:

    PATH=/usr/local/bin:/usr/bin:/bin
    if test "$HOME" != "/" ; then
    for dir in $HOME/bin/$CPU $HOME/bin $HOME/.local/bin/$CPU $HOME/.local/bin ; do
        test -d $dir && PATH=$dir:$PATH
    done
    fi

So, it is prudent to suggest that users check this before modifying
their .profile file. Furthermore, while `~/bin` has been informally used
for a while, `~/.local/bin` is part of the XDG Base Directory Spec,
which is increasingly embraced by the Linux ecosystem.